What we actually want to do is to share an object between handlers, so that: - every proxy has a dedicated handler - every handler has a dedicated handler model knowing the proxy - every  handler model has a client-defined model  that can be shared among them. The idea is that this client-defined model implements shared behavior acting on shared data (upon a proxified method execution).
Proxies are correctly mapped to handlers, handlers have a handler model linked to their proxy, and an optionnal model can be shared among all handlers to implement common behavior and hold shared data.

This is provided as  a usecase, the naming might be wrong, possibly the implementation could be better done.